# openapi.yaml
openapi: 3.0.0
info:
  title: task-api
  description: >-
    Task management REST API. Users own projects; projects contain tasks; access
    is scoped to the authenticated user at every layer.
  version: 0.1.0
components:
  schemas:
...
    def-1:
      description: A task
      type: object
      required:
        - id
        - projectId
        - title
        - description
        - status
        - priority
        - dueDate
        - position
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: string
          nullable: true
        position:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      title: taskResponseSchema
paths:
...
  /projects/{id}/tasks:
    get:
      parameters:
        - schema:
            type: string
            pattern: ^[0-9]{1,15}$
          in: path
          name: id
          required: true
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                description: A list of tasks
                type: array
                items:
                  $ref: '#/components/schemas/def-1'
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              additionalProperties: false
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  nullable: true
                status:
                  type: string
                  enum:
                    - todo
                    - in_progress
                    - done
                priority:
                  type: string
                  enum:
                    - low
                    - medium
                    - high
                dueDate:
                  type: string
                  format: date
                  nullable: true
                position:
                  type: integer
      parameters:
        - schema:
            type: string
            pattern: ^[0-9]{1,15}$
          in: path
          name: id
          required: true
      responses:
        '201':
          description: A task
          content:
            application/json:
              schema:
                description: A task
                type: object
                required:
                  - id
                  - projectId
                  - title
                  - description
                  - status
                  - priority
                  - dueDate
                  - position
                  - createdAt
                  - updatedAt
                properties:
                  id:
                    type: integer
                  projectId:
                    type: integer
                  title:
                    type: string
                  description:
                    type: string
                    nullable: true
                  status:
                    type: string
                  priority:
                    type: string
                  dueDate:
                    type: string
                    nullable: true
                  position:
                    type: integer
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
  /tasks/{id}:
    get:
      parameters:
        - schema:
            type: string
            pattern: ^[0-9]{1,15}$
          in: path
          name: id
          required: true
      responses:
        '200':
          description: A task
          content:
            application/json:
              schema:
                description: A task
                type: object
                required:
                  - id
                  - projectId
                  - title
                  - description
                  - status
                  - priority
                  - dueDate
                  - position
                  - createdAt
                  - updatedAt
                properties:
                  id:
                    type: integer
                  projectId:
                    type: integer
                  title:
                    type: string
                  description:
                    type: string
                    nullable: true
                  status:
                    type: string
                  priority:
                    type: string
                  dueDate:
                    type: string
                    nullable: true
                  position:
                    type: integer
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
    patch:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  nullable: true
                status:
                  type: string
                  enum:
                    - todo
                    - in_progress
                    - done
                priority:
                  type: string
                  enum:
                    - low
                    - medium
                    - high
                dueDate:
                  type: string
                  format: date
                  nullable: true
                position:
                  type: integer
      parameters:
        - schema:
            type: string
            pattern: ^[0-9]{1,15}$
          in: path
          name: id
          required: true
      responses:
        '200':
          description: A task
          content:
            application/json:
              schema:
                description: A task
                type: object
                required:
                  - id
                  - projectId
                  - title
                  - description
                  - status
                  - priority
                  - dueDate
                  - position
                  - createdAt
                  - updatedAt
                properties:
                  id:
                    type: integer
                  projectId:
                    type: integer
                  title:
                    type: string
                  description:
                    type: string
                    nullable: true
                  status:
                    type: string
                  priority:
                    type: string
                  dueDate:
                    type: string
                    nullable: true
                  position:
                    type: integer
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
    delete:
      parameters:
        - schema:
            type: string
            pattern: ^[0-9]{1,15}$
          in: path
          name: id
          required: true
      responses:
        '204':
          description: No content